requirements.txt
# Directory Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.10.1
pydantic-settings==2.6.1
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pyjwt==2.8.0
pytz==2023.3

jinja2==3.1.2

# Google OAuth Dependencies
authlib==1.2.1
httpx==0.25.2
email-validator==2.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
